# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simple Dictionary API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/token/validate": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Validate access token",
                "description": "Validates the current Bearer token or a provided user_token. Returns user data if valid.",
                "operationId": "validateToken",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_token": {
                                        "description": "Optional token to validate. If omitted, validates the Bearer token from Authorization header.",
                                        "type": "string",
                                        "example": "1|laravel_sanctum_token...",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token is valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "valid": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "john@example.com"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token is invalid or expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "valid": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token expired"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/matches/active": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get active match",
                "description": "Returns the currently active (in_progress) match for the authenticated user or guest",
                "operationId": "getActiveMatch",
                "parameters": [
                    {
                        "name": "guest_id",
                        "in": "query",
                        "description": "Guest identifier (if not authenticated)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Active match or null",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/Match"
                                                }
                                            ],
                                            "nullable": true,
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Match": {
                "description": "Match resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "language_from_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "language_to_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "dictionary_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "match_type": {
                        "type": "string",
                        "enum": [
                            "time",
                            "steps",
                            "race"
                        ],
                        "example": "time"
                    },
                    "match_type_params": {
                        "type": "object",
                        "example": "{\"duration\": 120}"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ],
                        "example": "in_progress"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-01T12:00:00Z",
                        "nullable": true
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "completion_reason": {
                        "type": "string",
                        "enum": [
                            "time_expired",
                            "steps_completed",
                            "all_players_left",
                            "forfeited",
                            "cancelled"
                        ],
                        "example": null,
                        "nullable": true
                    },
                    "participants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MatchUser"
                        }
                    },
                    "time_left": {
                        "description": "Seconds remaining (only during active time match)",
                        "type": "integer",
                        "example": 60,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "MatchUser": {
                "description": "Match participant resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 42,
                        "nullable": true
                    },
                    "guest_id": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "participant_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "participant_avatar": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg",
                        "nullable": true
                    },
                    "score": {
                        "type": "integer",
                        "example": 100
                    },
                    "answered_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "correct_answers_count": {
                        "type": "integer",
                        "example": 4
                    },
                    "place": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "is_winner": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_guest": {
                        "type": "boolean",
                        "example": false
                    },
                    "steps_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "skipped_steps_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "finished",
                            "spectating",
                            "left",
                            "disconnected",
                            "forfeited"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}